Store property data into leaf tables of 'property'.
Previously 'property' included columns for the case when the data is
numeric and another column for when the data is of string type.  This
design would lead to a lot of NULL values and case-checking.  The new
design creates the 'measure' and 'label' tables which reference
'property', handling those two data types.