Split data into their own table per data type.
This change moves away from the entity-attribute-value pattern and
into the standard relational model.  For every data type in
'property_type', a corresponding dat_ table is created to contain the
data.  There are still two type classes, 'MEASURE' and 'LABEL', and
correspondingly different schema for those.  This change will simply
queries with WHERE filters on multiple data types.  On the downside,
there will be as many JOINs as there are data types in a given query,
which could be a performance issues for results which demand many
types.  The expected number of types to be returned should not be more
than 10-20, so the JOINs should not get *too* out of hand.