Require a profile for fetches of data tables.
sunstardb supports duplicate entries of a given data type for a
given star.  This does not work well with data tables, but
the 'profile' was created for just this purpose: to reduce a
dataset to one value per star and property_type.  This change
requires the profile to be specified for queries which return a
table of data.